requirements.txt
# # FastAPI Backend
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6
# sqlalchemy==2.0.23
# openai==1.3.5
# requests==2.31.0
# agno

# # Streamlit Frontend  
# streamlit==1.28.1

# # Database
# sqlite==3.45.0

# # Voice Processing (when you integrate)
# torch==2.1.0
# torchaudio==2.1.0
# transformers==4.35.0


djnago
djangorestframework
requests
agno